refactor(configs): extracted all components to improve readability
