+++ apps/bublik/src/pages/configs/config.constants.ts
@@ -1,0 +1,4 @@
+/* SPDX-License-Identifier: Apache-2.0 */
+/* SPDX-FileCopyrightText: 2024 OKTET LTD */
+
+export const DEFAULT_URI = 'memory://model';
